reorg photometry module code
