+++ sbpy/photometry/__init__.py
@@ -3,4 +3,5 @@ sbpy Module for small body photometry
 """
 
 from .core import *
+from .iau import *
 from .bandpass import *
